Rework adapters with some inheritance.
- merge most of sqlite3 and psycopg driver
- make it easier to support other databases
