+++ aiosql/adapters/generic.py
@@ -1,71 +1,72 @@
 from contextlib import contextmanager
 
-from ..patterns import var_pattern
 
+class GenericAdapter:
+    """
+    Generic AioSQL Adapter suitable for `named` parameter style and no with support.
+    """
 
-def replacer(match):
-    gd = match.groupdict()
-    if gd["dblquote"] is not None:
-        return gd["dblquote"]
-    elif gd["quote"] is not None:
-        return gd["quote"]
-    else:
-        return f'{gd["lead"]}%({gd["var_name"]})s{gd["trail"]}'
-
-
-class PsycoPG2Adapter:
     @staticmethod
     def process_sql(_query_name, _op_type, sql):
-        return var_pattern.sub(replacer, sql)
+        return sql
 
     @staticmethod
     def select(conn, _query_name, sql, parameters, record_class=None):
-        with conn.cursor() as cur:
+        cur = conn.cursor()
+        try:
             cur.execute(sql, parameters)
             results = cur.fetchall()
             if record_class is not None:
-                column_names = [c.name for c in cur.description]
+                column_names = [c[0] for c in cur.description]
                 results = [record_class(**dict(zip(column_names, row))) for row in results]
+        finally:
+            cur.close()
         return results
 
     @staticmethod
     def select_one(conn, _query_name, sql, parameters, record_class=None):
-        with conn.cursor() as cur:
+        cur = conn.cursor()
+        try:
             cur.execute(sql, parameters)
             result = cur.fetchone()
             if result is not None and record_class is not None:
-                column_names = [c.name for c in cur.description]
+                column_names = [c[0] for c in cur.description]
                 result = record_class(**dict(zip(column_names, result)))
+        finally:
+            cur.close()
         return result
 
     @staticmethod
     @contextmanager
     def select_cursor(conn, _query_name, sql, parameters):
-        with conn.cursor() as cur:
-            cur.execute(sql, parameters)
+        cur = conn.cursor()
+        cur.execute(sql, parameters)
+        try:
             yield cur
+        finally:
+            cur.close()
 
     @staticmethod
     def insert_update_delete(conn, _query_name, sql, parameters):
-        with conn.cursor() as cur:
-            cur.execute(sql, parameters)
+        cur = conn.cursor()
+        cur.execute(sql, parameters)
+        cur.close()
 
     @staticmethod
     def insert_update_delete_many(conn, _query_name, sql, parameters):
-        with conn.cursor() as cur:
-            cur.executemany(sql, parameters)
+        cur = conn.cursor()
+        cur.executemany(sql, parameters)
+        cur.close()
 
     @staticmethod
     def insert_returning(conn, _query_name, sql, parameters):
         with conn.cursor() as cur:
             cur.execute(sql, parameters)
             res = cur.fetchone()
-            if res:
-                return res[0] if len(res) == 1 else res
-            else:
-                return None
+            return res[0] if res and len(res) == 1 else res
 
     @staticmethod
     def execute_script(conn, sql):
-        with conn.cursor() as cur:
-            cur.execute(sql)
+        cur = conn.cursor()
+        cur.execute(sql, parameters)
+        cur.close()
