+++ aiosql/adapters/psycopg.py
@@ -1,0 +1,4 @@
+from .pyformat import PyFormatAdapter
+
+class PsycoPGAdapter(PyFormatAdapter):
+    pass
