+++ aiosql/aiosql.py
@@ -3,8 +3,9 @@ from typing import Callable, Dict, Optional, Type, Union
 
 from .adapters.aiosqlite import AioSQLiteAdapter
 from .adapters.asyncpg import AsyncPGAdapter
-from .adapters.psycopg2 import PsycoPG2Adapter
-from .adapters.sqlite3 import SQLite3DriverAdapter
+from .adapters.psycopg import PsycoPGAdapter
+from .adapters.pyformat import PyFormatAdapter
+from .adapters.sqlite3 import SQLite3Adapter
 from .exceptions import SQLLoadException
 from .queries import Queries
 from .query_loader import QueryLoader
@@ -14,9 +15,9 @@ from .types import DriverAdapterProtocol
 _ADAPTERS: Dict[str, Callable[..., DriverAdapterProtocol]] = {
     "aiosqlite": AioSQLiteAdapter,
     "asyncpg": AsyncPGAdapter,
-    "psycopg2": PsycoPG2Adapter,
-    "psycopg": PsycoPG2Adapter,
-    "sqlite3": SQLite3DriverAdapter,
+    "psycopg": PsycoPGAdapter,
+    "psycopg2": PsycoPGAdapter,
+    "sqlite3": SQLite3Adapter,
 }
 
 
@@ -30,6 +31,15 @@ def _make_driver_adapter(
             driver_adapter = _ADAPTERS[driver_adapter]
         except KeyError:
             raise ValueError(f"Encountered unregistered driver_adapter: {driver_adapter}")
+    if isinstance(driver_adapter, Callable):
+        if hasattr(driver_adapter, 'paramstyle'):
+            style = getattr(driver_adapter, 'paramstyle')
+            if style == 'pyformat':
+                driver_adapter = PyFormatAdapter
+            elif style == 'named':
+                driver_adapter = GenericAdapter
+            else:
+                raise ValueError(f"Unexpected driver_adapter: {driver_adapter}")
 
     return driver_adapter()
